Implement agent architecture separation of concerns
  - Simplify BaseAgent to focus solely on prompt execution
  - Create ConductorAgent for orchestration decisions
  - Add AgentRunner to coordinate BaseAgent + ConductorAgent interactions
  - Refactor AgentRunner to use Run() method with BaseAgent as parameter
  - Add GetConductorAgent() and GetAgentRunner() factory methods
  - Fix TypeScript return type annotations in AgentRunner
  - Preserve original BaseAgent implementation in base-agent-original.ts

  This implements the clean separation discussed with Amith:
  - BaseAgent: Domain-specific execution only
  - ConductorAgent: Autonomous orchestration decisions
  - AgentRunner: Coordination and user interface
