+++ packages/AI/Agents/src/index.ts
@@ -1,2 +1,4 @@
 export * from './base-agent';
 export * from './agent-factory';
+export * from './conductor-agent';
+export * from './agent-runner';
